Support for DataMember attributes
In our project, we use DataMember attributes so that we don't need to deal with the __c ugliness of custom fields. The field name will be whatever we want and the datamember attribute will reflect the salesforce API name. The toolkit doesn't play well with this, since it is reflecting on the property names alone. The proposed code tries to handle this by looking for the datamember attribute first. If it is not available, it will pick up the property name.