add GetBlobAsync to IForceClient interface
https://github.com/developerforce/Force.com-Toolkit-for-NET/pull/156 added the method to ForceClient but didn't update the IForceClient interface.  It would be preferable to be able to keep using the interface even when needing to call GetBlobAsync.

The signature being added matches the signature that was added in that pull request.  I think using 'recordId' for the second parameter name would be more consistent with the other calls but changing the parameter name is a breaking change so I'm just adding the signature as it already exists.